Sensors-Process-Unit/src/requirements.txt
asyncio==3.4.3
certifi==2025.4.26
charset-normalizer==3.4.2
docopt==0.6.2
hbmqtt==0.9.6
idna==3.10
numpy
paho-mqtt==2.1.0
passlib==1.7.4
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
transitions==0.9.2
urllib3==2.4.0
paho-mqtt==2.1.0
websockets==15.0.1
